+++ 1.21/Common/src/main/java/fuzs/puzzleslib/api/entity/v1/DamageSourcesHelper.java
@@ -1,5 +1,6 @@
 package fuzs.puzzleslib.api.entity.v1;
 
+import fuzs.puzzleslib.api.init.v3.registry.LookupHelper;
 import net.minecraft.core.RegistryAccess;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceKey;
@@ -10,7 +11,8 @@ import net.minecraft.world.level.LevelReader;
 import org.jetbrains.annotations.Nullable;
 
 /**
- * A small helper class for creating new {@link DamageSource} instances, as vanilla's {@link net.minecraft.world.damagesource.DamageSources} does not allow for creating custom damage types.
+ * A small helper class for creating new {@link DamageSource} instances, as vanilla's
+ * {@link net.minecraft.world.damagesource.DamageSources} does not allow for creating custom damage types.
  */
 public final class DamageSourcesHelper {
 
@@ -34,7 +36,8 @@ public final class DamageSourcesHelper {
      *
      * @param level        registry access for retrieving dynamic {@link DamageType} registry
      * @param damageType   key for finding the {@link DamageType}
-     * @param directEntity the entity directly responsible for causing damage, like the mob attacking the player, or the arrow hitting the target
+     * @param directEntity the entity directly responsible for causing damage, like the mob attacking the player, or the
+     *                     arrow hitting the target
      * @return new {@link DamageSource} instance
      */
     public static DamageSource source(LevelReader level, ResourceKey<DamageType> damageType, @Nullable Entity directEntity) {
@@ -46,8 +49,10 @@ public final class DamageSourcesHelper {
      *
      * @param level         registry access for retrieving dynamic {@link DamageType} registry
      * @param damageType    key for finding the {@link DamageType}
-     * @param directEntity  the entity directly responsible for causing damage, like the mob attacking the player, or the arrow hitting the target
-     * @param causingEntity the entity that is responsible for <code>directEntity</code> to cause damage to the target, like the player that shot the arrow
+     * @param directEntity  the entity directly responsible for causing damage, like the mob attacking the player, or
+     *                      the arrow hitting the target
+     * @param causingEntity the entity that is responsible for <code>directEntity</code> to cause damage to the target,
+     *                      like the player that shot the arrow
      * @return new {@link DamageSource} instance
      */
     public static DamageSource source(LevelReader level, ResourceKey<DamageType> damageType, @Nullable Entity directEntity, @Nullable Entity causingEntity) {
@@ -59,11 +64,16 @@ public final class DamageSourcesHelper {
      *
      * @param registryAccess registry access for retrieving dynamic {@link DamageType} registry
      * @param damageType     key for finding the {@link DamageType}
-     * @param directEntity   the entity directly responsible for causing damage, like the mob attacking the player, or the arrow hitting the target
-     * @param causingEntity  the entity that is responsible for <code>directEntity</code> to cause damage to the target, like the player that shot the arrow
+     * @param directEntity   the entity directly responsible for causing damage, like the mob attacking the player, or
+     *                       the arrow hitting the target
+     * @param causingEntity  the entity that is responsible for <code>directEntity</code> to cause damage to the target,
+     *                       like the player that shot the arrow
      * @return new {@link DamageSource} instance
      */
     public static DamageSource source(RegistryAccess registryAccess, ResourceKey<DamageType> damageType, @Nullable Entity directEntity, @Nullable Entity causingEntity) {
-        return new DamageSource(registryAccess.registryOrThrow(Registries.DAMAGE_TYPE).getHolderOrThrow(damageType), directEntity, causingEntity);
+        return new DamageSource(LookupHelper.lookup(registryAccess, Registries.DAMAGE_TYPE, damageType),
+                directEntity,
+                causingEntity
+        );
     }
 }