+++ 1.21/Common/src/main/java/fuzs/puzzleslib/api/init/v3/registry/RegistryFactory.java
@@ -18,6 +18,21 @@ public interface RegistryFactory {
      * Create and register a {@link net.minecraft.core.MappedRegistry}.
      * <p>
      * Calls {@link #register(Registry)} as part of the implementation.
+     * <p>
+     * Registry contents are synced to clients.
+     *
+     * @param registryKey the registry key
+     * @param <T>         registry values type
+     * @return the new registry
+     */
+    default <T> Registry<T> create(ResourceKey<Registry<T>> registryKey) {
+        return this.create(registryKey, true);
+    }
+
+    /**
+     * Create and register a {@link net.minecraft.core.MappedRegistry}.
+     * <p>
+     * Calls {@link #register(Registry)} as part of the implementation.
      *
      * @param registryKey the registry key
      * @param synced      sync registry contents to clients, so that numeric registry ids can be used in networking
@@ -25,6 +40,37 @@ public interface RegistryFactory {
      * @return the new registry
      */
     <T> Registry<T> create(ResourceKey<Registry<T>> registryKey, boolean synced);
+
+    /**
+     * Create and register a {@link net.minecraft.core.DefaultedMappedRegistry}.
+     * <p>
+     * Calls {@link #register(Registry)} as part of the implementation.
+     * <p>
+     * Registry contents are synced to clients.
+     *
+     * @param registryKey the registry key
+     * @param defaultKey  the default value key
+     * @param <T>         registry values type
+     * @return the new registry
+     */
+    default <T> Registry<T> create(ResourceKey<Registry<T>> registryKey, String defaultKey) {
+        return this.create(registryKey, defaultKey, true);
+    }
+
+    /**
+     * Create and register a {@link net.minecraft.core.DefaultedMappedRegistry}.
+     * <p>
+     * Calls {@link #register(Registry)} as part of the implementation.
+     *
+     * @param registryKey the registry key
+     * @param defaultKey  the default value key
+     * @param synced      sync registry contents to clients, so that numeric registry ids can be used in networking
+     * @param <T>         registry values type
+     * @return the new registry
+     */
+    default <T> Registry<T> create(ResourceKey<Registry<T>> registryKey, String defaultKey, boolean synced) {
+        return this.create(registryKey, registryKey.location().withPath(defaultKey), synced);
+    }
 
     /**
      * Create and register a {@link net.minecraft.core.DefaultedMappedRegistry}.